Extracted subcomponent CustomDot
